Seamless component Metadata across tools (#10863)
* Renamed components for polarion

* Relam and Convert2RHEL tests merged

* ContentManagement Tests conversion
