+++ tests/foreman/api/test_location.py
@@ -9,7 +9,7 @@ http://theforeman.org/api/apidoc/v2/locations.html
 
 :CaseLevel: Acceptance
 
-:CaseComponent: OrganizationsLocations
+:CaseComponent: OrganizationsandLocations
 
 :Team: Endeavour
