+++ tests/foreman/api/test_repository.py
@@ -38,6 +38,7 @@ from robottelo.api.utils import promote
 from robottelo.config import settings
 from robottelo.constants import DataFile
 from robottelo.constants import repos as repo_constants
+from robottelo.content_info import get_repo_files_by_url
 from robottelo.logging import logger
 from robottelo.utils import datafactory
 from robottelo.utils.datafactory import parametrized
@@ -1286,10 +1287,86 @@ class TestRepository:
         command_output = target_sat.execute('foreman-rake katello:correct_repositories COMMIT=true')
         assert 'Recreating' in command_output.stdout and 'TaskError' not in command_output.stdout
 
+    @pytest.mark.tier2
+    def test_positive_mirroring_policy(self, target_sat):
+        """Assert that the content of a repository with 'Mirror Policy' enabled
+        is restored properly after resync.
+
+        :id: cbf1c781-cb96-4b4a-bae2-15c9f5be5e50
+
+        :steps:
+            1. Create and sync a repo with 'Mirror Policy - mirror complete' enabled.
+            2. Remove all packages from the repo and upload another one.
+            3. Resync the repo again.
+            4. Check the content was restored properly.
+
+        :expectedresults:
+            1. The resync restores the original content properly.
+
+        :CaseLevel: System
+        """
+        repo_url = settings.repos.yum_0.url
+        packages_count = constants.FAKE_0_YUM_REPO_PACKAGES_COUNT
+
+        org = entities.Organization().create()
+        prod = entities.Product(organization=org).create()
+        repo = entities.Repository(
+            download_policy='immediate',
+            mirroring_policy='mirror_complete',
+            product=prod,
+            url=repo_url,
+        ).create()
+        repo.sync()
+        repo = repo.read()
+        assert repo.content_counts['rpm'] == packages_count
+
+        # remove all packages from the repo and upload another one
+        packages = entities.Package(repository=repo).search(query={'per_page': '1000'})
+        repo.remove_content(data={'ids': [package.id for package in packages]})
+
+        with open(DataFile.RPM_TO_UPLOAD, 'rb') as handle:
+            repo.upload_content(files={'content': handle})
+
+        repo = repo.read()
+        assert repo.content_counts['rpm'] == 1
+        files = get_repo_files_by_url(repo.full_path)
+        assert len(files) == 1
+        assert constants.RPM_TO_UPLOAD in files
+
+        # resync the repo again and check the content
+        repo.sync()
+
+        repo = repo.read()
+        assert repo.content_counts['rpm'] == packages_count
+        files = get_repo_files_by_url(repo.full_path)
+        assert len(files) == packages_count
+        assert constants.RPM_TO_UPLOAD not in files
+
 
 @pytest.mark.run_in_one_thread
 class TestRepositorySync:
     """Tests for ``/katello/api/repositories/:id/sync``."""
+
+    @pytest.mark.tier2
+    def test_positive_sync_repos_with_lots_files(self):
+        """Attempt to synchronize repository containing a lot of files inside
+        rpms.
+
+        :id: 2cc09ce3-d5df-4caa-956a-78f83a7735ca
+
+        :customerscenario: true
+
+        :BZ: 1404345
+
+        :CaseLevel: Integration
+
+        :expectedresults: repository was successfully synchronized
+        """
+        org = entities.Organization().create()
+        product = entities.Product(organization=org).create()
+        repo = entities.Repository(product=product, url=settings.repos.yum_8.url).create()
+        response = repo.sync()
+        assert response, f"Repository {repo} failed to sync."
 
     @pytest.mark.tier2
     def test_positive_sync_rh(self, module_entitlement_manifest_org):
@@ -1474,6 +1551,49 @@ class TestRepositorySync:
         )
         assert prod_log_out.status == 0
         assert "(0 rows)" in prod_log_out.stdout
+
+    @pytest.mark.parametrize(
+        'distro',
+        [
+            ver
+            for ver in settings.supportability.content_hosts.rhel.versions
+            if isinstance(ver, int)
+        ],
+    )
+    def test_positive_sync_kickstart_check_os(
+        self, module_entitlement_manifest_org, distro, target_sat
+    ):
+        """Sync rhel KS repo and assert that OS was created
+
+        :id: f84bcf1b-717e-40e7-82ee-000eead45249
+
+        :Parametrized: Yes
+
+        :steps:
+            1. Enable and sync a kickstart repo.
+            2. Check that OS with corresponding version.
+
+        :expectedresults:
+            1. OS with corresponding version was created.
+
+        """
+        distro = f'rhel{distro} + "_bos"' if distro > 7 else f'rhel{distro}'
+        repo_id = target_sat.api_factory.enable_rhrepo_and_fetchid(
+            basearch='x86_64',
+            org_id=module_entitlement_manifest_org.id,
+            product=constants.REPOS['kickstart'][distro]['product'],
+            reposet=constants.REPOSET['kickstart'][distro],
+            repo=constants.REPOS['kickstart'][distro]['name'],
+            releasever=constants.REPOS['kickstart'][distro]['version'],
+        )
+        rh_repo = entities.Repository(id=repo_id).read()
+        rh_repo.sync()
+
+        major, minor = constants.REPOS['kickstart'][distro]['version'].split('.')
+        os = entities.OperatingSystem().search(
+            query={'search': f'name="RedHat" AND major="{major}" AND minor="{minor}"'}
+        )
+        assert len(os)
 
 
 class TestDockerRepository:
