+++ tests/foreman/cli/test_ostreebranch.py
@@ -1,12 +1,12 @@
 """Test class for Ostree Branch CLI.
 
-:Requirement: Ostreebranch
+:Requirement: Repositories
 
 :CaseAutomation: Automated
 
 :CaseLevel: Acceptance
 
-:CaseComponent: ContentManagement
+:CaseComponent: Repositories
 
 :Team: Phoenix
