+++ tests/foreman/destructive/test_katello_agent.py
@@ -6,7 +6,7 @@
 
 :CaseLevel: Component
 
-:CaseComponent: Katello-agent
+:CaseComponent: katello-agent
 
 :Team: Phoenix
