+++ tests/foreman/ui/test_modulestreams.py
@@ -1,12 +1,12 @@
 """Test class for module_streams UI
 
-:Requirement: module_streams
+:Requirement: Repositories
 
 :CaseAutomation: Automated
 
 :CaseLevel: Acceptance
 
-:CaseComponent: ContentManagement
+:CaseComponent: Repositories
 
 :Team: Phoenix
