+++ tests/foreman/ui/test_package.py
@@ -1,12 +1,12 @@
 """Test class for Packages UI
 
-:Requirement: Packages
+:Requirement: Repositories
 
 :CaseAutomation: Automated
 
 :CaseLevel: Component
 
-:CaseComponent: ContentManagement
+:CaseComponent: Repositories
 
 :Team: Phoenix
